Improvements to my code as suggested by foosel.
Using Smartplug as object instead of calling it with subprocess.
Resuming a print will only turn on the powerplug if a certain layer was reached before.
Default values for settings are now environment unspecific.
Reacting to M106/M107 Gcodes is now toggleable in settings.
Fan on will now be triggered by M106 S>0. Fan off will be triggered by M106 S0 or M107